Install Adobe Acrobat Reader DC
After the AdobeReader_dc_en_ta_install.dmg is dropped into the /private/tmp from a .dmg wrapper. It mounts the install dmg and runs the Adobe Acrobat Reader DC Install app. It then checks to see if the process is running. Once the process stops running it dismounts the install dmg.